Move tests related to Frame repr into a separate file
